requirements/base.txt
# Bleeding edge Django
django==1.8.2

# Configuration
django-environ==0.3.0
django-secure==1.0.1



# Forms
django-braces==1.8.0
django-crispy-forms==1.4.0
django-floppyforms==1.4.1

# Models
django-model-utils==2.2

# Images
Pillow==2.8.2

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.20.0

# For the persistence stores
MySQL-python==1.2.5

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.8.0

# Time zones support
pytz==2015.4

# Your custom requirements go here
celery==3.1.18
# Smartmin for django 1.8
git+git://github.com/nyaruka/smartmin.git@995361b21491018c8d9d604b977158d0ba63ea20
django-phonenumber-field==0.7.2
